refactor(domain): domain 패키지 재구성
