+++ src/main/java/domain/movement/strategy/MovementStrategy.java
@@ -1,6 +1,7 @@
-package domain.strategy;
+package domain.movement.strategy;
 
-import domain.game.Position;
+import domain.common.Position;
+import domain.movement.Path;
 import java.util.List;
 
 public interface MovementStrategy {
