+++ src/main/java/domain/movement/strategy/PalaceGeneralGuardStrategy.java
@@ -1,8 +1,10 @@
-package domain.strategy;
+package domain.movement.strategy;
 
-import domain.game.Position;
-import domain.game.Side;
-import domain.game.Palace;
+import domain.common.Direction;
+import domain.common.Position;
+import domain.common.Side;
+import domain.board.Palace;
+import domain.movement.Path;
 import java.util.ArrayList;
 import java.util.List;
