+++ src/main/java/domain/movement/strategy/PalaceSoldierDiagonalStrategy.java
@@ -1,8 +1,9 @@
-package domain.strategy;
+package domain.movement.strategy;
 
-import domain.game.Position;
-import domain.game.Side;
-import domain.game.Palace;
+import domain.common.Position;
+import domain.common.Side;
+import domain.board.Palace;
+import domain.movement.Path;
 import java.util.List;
 
 public class PalaceSoldierDiagonalStrategy implements MovementStrategy {
