+++ src/main/java/domain/movement/strategy/SequenceStrategy.java
@@ -1,6 +1,8 @@
-package domain.strategy;
+package domain.movement.strategy;
 
-import domain.game.Position;
+import domain.common.Direction;
+import domain.common.Position;
+import domain.movement.Path;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
