+++ src/main/java/domain/piece/Guard.java
@@ -1,10 +1,10 @@
 package domain.piece;
 
-import domain.game.Position;
-import domain.game.Side;
+import domain.common.Position;
+import domain.common.Side;
 import domain.board.BoardReader;
-import domain.strategy.MovementStrategy;
-import domain.strategy.Path;
+import domain.movement.strategy.MovementStrategy;
+import domain.movement.Path;
 import java.util.List;
 
 public class Guard extends Piece {
