+++ src/main/java/domain/turn/ActiveTurn.java
@@ -1,4 +1,4 @@
-package domain.state;
+package domain.turn;
 
 public class ActiveTurn implements TurnState {
