+++ src/main/java/domain/turn/InactiveTurn.java
@@ -1,4 +1,4 @@
-package domain.state;
+package domain.turn;
 
 public class InactiveTurn implements TurnState {
