+++ src/test/java/domain/common/PositionTest.java
@@ -1,4 +1,4 @@
-package domain.game;
+package domain.common;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
